+++ cheetah-raft/src/main/java/raft/protocol/request/AddRequest.java
@@ -16,7 +16,7 @@ public class AddRequest extends BaseRequest implements Serializable {
     //领导人的任期号
     private int term;
     //领导人的 Id，以便于跟随者重定向请求
-    private int leaderId;
+    private long leaderId;
     //新的日志条目紧随之前的索引值, from -1
     private long prevLogIndex;
     //prevLogIndex 条目的任期号
@@ -26,9 +26,9 @@ public class AddRequest extends BaseRequest implements Serializable {
     //log entry
     private List<RaftLog.LogEntry> logEntries;
 
-    private int serverId;
+    private long serverId;
 
-    public AddRequest (int term, int leaderId, long prevLogIndex,
+    public AddRequest (int term, long leaderId, long prevLogIndex,
                        int prevLogTerm, long leaderCommit) {
         this.term = term;
         this.leaderId = leaderId;
@@ -37,7 +37,7 @@ public class AddRequest extends BaseRequest implements Serializable {
         this.leaderCommit = leaderCommit;
         logEntries = new ArrayList<>();
     }
-    public AddRequest (int term, int leaderId, long prevLogIndex,
+    public AddRequest (int term, long leaderId, long prevLogIndex,
                        int prevLogTerm, long leaderCommit, List<RaftLog.LogEntry> logEntries) {
         this.term = term;
         this.leaderId = leaderId;
@@ -48,7 +48,7 @@ public class AddRequest extends BaseRequest implements Serializable {
     }
 
     public void setAddress(String localHost, int localPort,
-                           String remoteHost, int remotePort, int serverId) {
+                           String remoteHost, int remotePort, long serverId) {
         this.localHost = localHost;
         this.localPort = localPort;
         this.remoteHost = remoteHost;
@@ -64,11 +64,11 @@ public class AddRequest extends BaseRequest implements Serializable {
         this.term = term;
     }
 
-    public int getLeaderId() {
+    public long getLeaderId() {
         return leaderId;
     }
 
-    public void setLeaderId(int leaderId) {
+    public void setLeaderId(long leaderId) {
         this.leaderId = leaderId;
     }
 
@@ -105,11 +105,11 @@ public class AddRequest extends BaseRequest implements Serializable {
         this.logEntries = logEntries;
     }
 
-    public int getServerId() {
+    public long getServerId() {
         return serverId;
     }
 
-    public void setServerId(int serverId) {
+    public void setServerId(long serverId) {
         this.serverId = serverId;
     }
 }
