+++ cheetah-raft/src/main/java/raft/protocol/request/GetLeaderRequest.java
@@ -11,14 +11,14 @@ import java.io.Serializable;
  */
 public class GetLeaderRequest extends BaseRequest implements Serializable {
 
-    private int serverId;
+    private long serverId;
 
     public GetLeaderRequest () {
 
     }
 
     public void setAddress(String localHost, int localPort,
-                           String remoteHost, int remotePort, int serverId) {
+                           String remoteHost, int remotePort, long serverId) {
         this.localHost = localHost;
         this.localPort = localPort;
         this.remoteHost = remoteHost;
@@ -26,11 +26,11 @@ public class GetLeaderRequest extends BaseRequest implements Serializable {
         this.serverId = serverId;
     }
 
-    public int getServerId() {
+    public long getServerId() {
         return serverId;
     }
 
-    public void setServerId(int serverId) {
+    public void setServerId(long serverId) {
         this.serverId = serverId;
     }
 }
