+++ cheetah-raft/src/main/java/raft/protocol/request/SetKVRequest.java
@@ -10,14 +10,14 @@ import raft.model.BaseRequest;
 public class SetKVRequest extends BaseRequest {
 
     private String setCommand;
-    private int serverId;
+    private long serverId;
 
     public SetKVRequest (String setCommand) {
         this.setCommand = setCommand;
     }
 
     public void setAddress(String localHost, int localPort,
-                           String remoteHost, int remotePort, int serverId) {
+                           String remoteHost, int remotePort, long serverId) {
         this.localHost = localHost;
         this.localPort = localPort;
         this.remoteHost = remoteHost;
@@ -33,11 +33,11 @@ public class SetKVRequest extends BaseRequest {
         this.setCommand = setCommand;
     }
 
-    public int getServerId() {
+    public long getServerId() {
         return serverId;
     }
 
-    public void setServerId(int serverId) {
+    public void setServerId(long serverId) {
         this.serverId = serverId;
     }
 }
