+++ cheetah-raft/src/main/java/raft/protocol/response/CommandExecuteResponse.java
@@ -9,11 +9,11 @@ public class CommandExecuteResponse extends RaftResponse {
 
     private String commandRtr;
 
-    public CommandExecuteResponse (int serverId) {
+    public CommandExecuteResponse (long serverId) {
         super(serverId);
     }
 
-    public CommandExecuteResponse(int serverId, String commandRtr) {
+    public CommandExecuteResponse(long serverId, String commandRtr) {
         super(serverId);
         this.commandRtr = commandRtr;
     }
