+++ cheetah-raft/src/main/java/raft/protocol/response/SetKVResponse.java
@@ -9,11 +9,11 @@ public class SetKVResponse extends RaftResponse{
 
     private String respMessage;
 
-    public SetKVResponse(int serverId) {
+    public SetKVResponse(long serverId) {
         super(serverId);
     }
 
-    public SetKVResponse(int serverId, String respMessage) {
+    public SetKVResponse(long serverId, String respMessage) {
         super(serverId);
         this.respMessage = respMessage;
     }
